+++ examples/src/main/kotlin/space/kscience/kmath/tensors/multik.kt
@@ -7,12 +7,12 @@ package space.kscience.kmath.tensors
 
 import org.jetbrains.kotlinx.multik.api.Multik
 import org.jetbrains.kotlinx.multik.api.ndarray
-import space.kscience.kmath.multik.multikND
+import space.kscience.kmath.multik.multikAlgebra
 import space.kscience.kmath.nd.one
 import space.kscience.kmath.operations.DoubleField
 
-fun main(): Unit = with(DoubleField.multikND) {
+fun main(): Unit = with(DoubleField.multikAlgebra) {
     val a = Multik.ndarray(intArrayOf(1, 2, 3)).asType<Double>().wrap()
     val b = Multik.ndarray(doubleArrayOf(1.0, 2.0, 3.0)).wrap()
-    one(a.shape) - a + b * 3
+    one(a.shape) - a + b * 3.0
 }