+++ kmath-tensors/src/commonMain/kotlin/space/kscience/kmath/tensors/api/AnalyticTensorAlgebra.kt
@@ -5,6 +5,7 @@
 
 package space.kscience.kmath.tensors.api
 
+import space.kscience.kmath.nd.StructureND
 import space.kscience.kmath.operations.Field
 
 
@@ -18,7 +19,7 @@ public interface AnalyticTensorAlgebra<T, A : Field<T>> : TensorPartialDivisionA
     /**
      * @return the mean of all elements in the input tensor.
      */
-    public fun Tensor<T>.mean(): T
+    public fun StructureND<T>.mean(): T
 
     /**
      * Returns the mean of each row of the input tensor in the given dimension [dim].
@@ -31,12 +32,12 @@ public interface AnalyticTensorAlgebra<T, A : Field<T>> : TensorPartialDivisionA
      * @param keepDim whether the output tensor has [dim] retained or not.
      * @return the mean of each row of the input tensor in the given dimension [dim].
      */
-    public fun Tensor<T>.mean(dim: Int, keepDim: Boolean): Tensor<T>
+    public fun StructureND<T>.mean(dim: Int, keepDim: Boolean): Tensor<T>
 
     /**
      * @return the standard deviation of all elements in the input tensor.
      */
-    public fun Tensor<T>.std(): T
+    public fun StructureND<T>.std(): T
 
     /**
      * Returns the standard deviation of each row of the input tensor in the given dimension [dim].
@@ -49,12 +50,12 @@ public interface AnalyticTensorAlgebra<T, A : Field<T>> : TensorPartialDivisionA
      * @param keepDim whether the output tensor has [dim] retained or not.
      * @return the standard deviation of each row of the input tensor in the given dimension [dim].
      */
-    public fun Tensor<T>.std(dim: Int, keepDim: Boolean): Tensor<T>
+    public fun StructureND<T>.std(dim: Int, keepDim: Boolean): Tensor<T>
 
     /**
      * @return the variance of all elements in the input tensor.
      */
-    public fun Tensor<T>.variance(): T
+    public fun StructureND<T>.variance(): T
 
     /**
      * Returns the variance of each row of the input tensor in the given dimension [dim].
@@ -67,57 +68,57 @@ public interface AnalyticTensorAlgebra<T, A : Field<T>> : TensorPartialDivisionA
      * @param keepDim whether the output tensor has [dim] retained or not.
      * @return the variance of each row of the input tensor in the given dimension [dim].
      */
-    public fun Tensor<T>.variance(dim: Int, keepDim: Boolean): Tensor<T>
+    public fun StructureND<T>.variance(dim: Int, keepDim: Boolean): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.exp.html
-    public fun Tensor<T>.exp(): Tensor<T>
+    public fun StructureND<T>.exp(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.log.html
-    public fun Tensor<T>.ln(): Tensor<T>
+    public fun StructureND<T>.ln(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.sqrt.html
-    public fun Tensor<T>.sqrt(): Tensor<T>
+    public fun StructureND<T>.sqrt(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.acos.html#torch.cos
-    public fun Tensor<T>.cos(): Tensor<T>
+    public fun StructureND<T>.cos(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.acos.html#torch.acos
-    public fun Tensor<T>.acos(): Tensor<T>
+    public fun StructureND<T>.acos(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.acosh.html#torch.cosh
-    public fun Tensor<T>.cosh(): Tensor<T>
+    public fun StructureND<T>.cosh(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.acosh.html#torch.acosh
-    public fun Tensor<T>.acosh(): Tensor<T>
+    public fun StructureND<T>.acosh(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.asin.html#torch.sin
-    public fun Tensor<T>.sin(): Tensor<T>
+    public fun StructureND<T>.sin(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.asin.html#torch.asin
-    public fun Tensor<T>.asin(): Tensor<T>
+    public fun StructureND<T>.asin(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.asin.html#torch.sinh
-    public fun Tensor<T>.sinh(): Tensor<T>
+    public fun StructureND<T>.sinh(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.asin.html#torch.asinh
-    public fun Tensor<T>.asinh(): Tensor<T>
+    public fun StructureND<T>.asinh(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.atan.html#torch.tan
-    public fun Tensor<T>.tan(): Tensor<T>
+    public fun StructureND<T>.tan(): Tensor<T>
 
     //https://pytorch.org/docs/stable/generated/torch.atan.html#torch.atan
-    public fun Tensor<T>.atan(): Tensor<T>
+    public fun StructureND<T>.atan(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.atanh.html#torch.tanh
-    public fun Tensor<T>.tanh(): Tensor<T>
+    public fun StructureND<T>.tanh(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.atanh.html#torch.atanh
-    public fun Tensor<T>.atanh(): Tensor<T>
+    public fun StructureND<T>.atanh(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.ceil.html#torch.ceil
-    public fun Tensor<T>.ceil(): Tensor<T>
+    public fun StructureND<T>.ceil(): Tensor<T>
 
     //For information: https://pytorch.org/docs/stable/generated/torch.floor.html#torch.floor
-    public fun Tensor<T>.floor(): Tensor<T>
+    public fun StructureND<T>.floor(): Tensor<T>
 
 }