+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - Buffer algebra does not require size anymore
 - Operations -> Ops
 - Default Buffer and ND algebras are now Ops and lack neutral elements (0, 1) as well as algebra-level shapes.
+- Tensor algebra takes read-only structures as input and inherits AlgebraND
 
 ### Deprecated
 - Specialized `DoubleBufferAlgebra`